+++ lib/raw.dart
@@ -8,6 +8,7 @@ export 'src/raw/picker_builders.dart';
 export 'src/raw/picker_config.dart';
 export 'src/raw/picker_debug.dart';
 export 'src/raw/search_anchor_picker.dart';
+export 'src/raw/widgets/default_picker_field_trigger.dart';
 export 'src/raw/widgets/overflow_tooltip_text.dart';
 export 'src/raw/widgets/passive_tooltip.dart';
 export 'src/raw/widgets/picker_defaults.dart';